backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
tensorflow==2.16.1
scikit-learn==1.5.0
joblib==1.4.2
opencv-python-headless==4.10.0.84
Pillow==10.3.0
numpy==1.26.4
httpx==0.27.0
requests==2.32.3
sendgrid==6.11.0
jinja2==3.1.4
pydantic[email]==2.7.4
alembic==1.13.1
